:hammer: refactor, template for session ui.
